Add files via upload
When you need to communicate with your server, you simply uncover the concealed series of NKK Smartswitches. You scroll through the displayed options until you find your desired item. Select your choice and the waiter is notified and brings it in a few minutes. The product features a unique LED display which illuminates various meal options ranging from additional utensils and napkins to ordering additional food or drink. The device communicates with a server display which notifies the table and request of a customer.